# bookingRoom-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Abdellah",
    "description": "Documentation OpenApi pour Spring Security",
    "termsOfService": "Conditions d'utilisation",
    "contact": {
      "name": "Abdellah",
      "url": "https://portfoliokaba.com",
      "email": "abdallahkaba98@gmail.com"
    },
    "license": {
      "name": "Nom de la licence",
      "url": "https://some-url.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Local ENV"
    },
    {
      "url": "https://portfoliokaba.com",
      "description": "PROD ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/booking/to-book": {
      "post": {
        "tags": [
          "Booking"
        ],
        "operationId": "saveBooking",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/booking/get-by-confirmation-code/{confirmation-code}": {
      "get": {
        "tags": [
          "Booking"
        ],
        "operationId": "findBookingByConfirmationCode",
        "parameters": [
          {
            "name": "confirmation-code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/booking/all": {
      "get": {
        "tags": [
          "Booking"
        ],
        "operationId": "getAllBookings",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseBookingResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/booking/cancel/{booking-id}": {
      "delete": {
        "tags": [
          "Booking"
        ],
        "operationId": "cancelBooking",
        "parameters": [
          {
            "name": "booking-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BookingRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "checkInDate": {
            "type": "string",
            "format": "date"
          },
          "checkOutDate": {
            "type": "string",
            "format": "date"
          },
          "numOfAdults": {
            "type": "integer",
            "format": "int32"
          },
          "numOfChildren": {
            "type": "integer",
            "format": "int32"
          },
          "totalNumOfGuest": {
            "type": "integer",
            "format": "int32"
          },
          "bookingConfirmationCode": {
            "type": "string"
          },
          "roomId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "BookingResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "checkInDate": {
            "type": "string",
            "format": "date"
          },
          "checkOutDate": {
            "type": "string",
            "format": "date"
          },
          "numOfAdults": {
            "type": "integer",
            "format": "int32"
          },
          "numOfChildren": {
            "type": "integer",
            "format": "int32"
          },
          "totalNumOfGuest": {
            "type": "integer",
            "format": "int32"
          },
          "bookingConfirmationCode": {
            "type": "string"
          },
          "roomType": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          }
        }
      },
      "PageResponseBookingResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookingResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Description de l'authentification JWT",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}